requirements.txt
africastalking==1.1.9
amqp==2.5.1
Babel==2.7.0
beautifulsoup4==4.8.1
billiard==3.6.1.0
bootstrap4==0.1.0
bs4==0.0.1
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.3.0
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
config==0.4.2
contextlib2==0.5.5
cssselect2==0.2.2
defusedxml==0.6.0
Django==2.2.6
django-advanced-reports==0.9.24
django-allauth==0.40.0
django-background-tasks==1.2.0
django-bootstrap3==11.1.0
django-bootstrap4==1.0.1
django-celery-beat==1.5.0
django-compat==1.0.15
django-crispy-forms==1.8.0
django-filter==2.2.0
django-getenv==1.3.2
django-report-builder==6.3.2
django-smart-selects-fix==1.0
django-timezone-field==3.1
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
dnspython==1.16.0
docopt==0.6.2
et-xmlfile==1.0.1
eventlet==0.25.1
flower==0.9.3
greenlet==0.4.15
html5lib==1.0.1
idna==2.8
importlib-metadata==0.23
jdcal==1.4.1
kombu==4.6.5
monotonic==1.5
more-itertools==7.2.0
num2words==0.5.10
oauthlib==3.1.0
openpyxl==3.0.0
Pillow==6.2.1
psycopg2==2.8.4
pycparser==2.19
Pygments==2.4.2
Pyphen==0.9.5
python-crontab==2.4.0
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.3
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.2.0
schema==0.7.1
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
tinycss2==1.0.2
tornado==5.1.1
urllib3==1.25.6
vine==1.3.0
WeasyPrint==50
webencodings==0.5.1
zipp==0.6.0
